+++ src/squlearn/__init__.py
@@ -1,1 +1,14 @@
+"""A library for quantum machine learning following the sklearn standard."""
+from .util import Executor
+
 __version__ = "0.1.0"
+
+__all__ = [
+    "Executor",
+    "expectation_operator",
+    "feature_map",
+    "kernel",
+    "optimizers",
+    "qnn",
+    "util",
+]
