+++ src/squlearn/expectation_operator/__init__.py
@@ -1,6 +1,3 @@
-from .expectation_operator_base import ExpectationOperatorBase
-from .expectation_operator_derivatives import ExpectationOperatorDerivatives
-
 from .expectation_operator_implemented.summed_probabilities import SummedProbabilities
 from .expectation_operator_implemented.summed_paulis import SummedPaulis
 from .expectation_operator_implemented.single_pauli import SinglePauli
@@ -11,8 +8,6 @@ from .expectation_operator_implemented.custom_expectation_operator import (
 from .expectation_operator_implemented.ising_hamiltonian import IsingHamiltonian
 
 __all__ = [
-    "ExpectationOperatorBase",
-    "ExpectationOperatorDerivatives",
     "SinglePauli",
     "SummedPaulis",
     "SingleProbability",
