+++ src/squlearn/kernel/matrix/__init__.py
@@ -1,5 +1,4 @@
-from .kernel_matrix_base import KernelMatrixBase
 from .fidelity_kernel import FidelityKernel
 from .projected_quantum_kernel import ProjectedQuantumKernel
 
-__all__ = ["KernelMatrixBase", "FidelityKernel", "ProjectedQuantumKernel"]
+__all__ = ["FidelityKernel", "ProjectedQuantumKernel"]
