+++ src/squlearn/kernel/ml/__init__.py
@@ -1,4 +1,3 @@
-from .kernel_ml_base import KernelMlBase
 from .qgpr import QGPR
 from .qgpc import QGPC
 from .qkrr import QKRR
@@ -6,4 +5,4 @@ from .qsvr import QSVR
 from .qsvc import QSVC
 
 
-__all__ = ["KernelMlBase", "QGPR", "QKRR", "QGPC", "QSVR", "QSVC"]
+__all__ = ["QGPR", "QKRR", "QGPC", "QSVR", "QSVC"]
