+++ src/squlearn/kernel/ml/kernel_util.py
@@ -2,7 +2,7 @@
 import numpy as np
 import scipy
 from sklearn.gaussian_process.kernels import Kernel
-from ..matrix import KernelMatrixBase
+from ..matrix.kernel_matrix_base import KernelMatrixBase
 
 
 def kernel_wrapper(kernel_matrix: KernelMatrixBase):
