+++ src/squlearn/kernel/ml/qgpc.py
@@ -1,6 +1,6 @@
 # QGPC
 import numpy as np
-from ..matrix import KernelMatrixBase
+from ..matrix.kernel_matrix_base import KernelMatrixBase
 from .kernel_util import kernel_wrapper
 from sklearn.gaussian_process import GaussianProcessClassifier
