+++ src/squlearn/kernel/ml/qsvr.py
@@ -1,4 +1,4 @@
-from ..matrix import KernelMatrixBase
+from ..matrix.kernel_matrix_base import KernelMatrixBase
 
 from sklearn.svm import SVR
