+++ src/squlearn/kernel/optimization/__init__.py
@@ -1,13 +1,5 @@
-from .kernel_optimization_base import KernelOptimizerBase
-from .kernel_loss_base import KernelLossBase
 from .kernel_optimizer import KernelOptimizer
 from .negative_log_likelihood import NLL
 from .target_alignment import TargetAlignment
 
-__all__ = [
-    "KernelOptimizerBase",
-    "KernelOptimizer",
-    "KernelLossBase",
-    "NLL",
-    "TargetAlignment",
-]
+__all__ = ["KernelOptimizer", "NLL", "TargetAlignment"]
