+++ src/squlearn/qnn/__init__.py
@@ -1,3 +1,7 @@
-from .qnn import QNN
+"""QNN module for classification and regression."""
+from .loss import SquaredLoss, VarianceLoss
+from .qnnc import QNNClassifier
+from .qnnr import QNNRegressor
+from .training import get_variance_fac
 
-__all__ = ["QNN"]
+__all__ = ["SquaredLoss", "VarianceLoss", "QNNClassifier", "QNNRegressor", "get_variance_fac"]
