+++ src/squlearn/qnn/qnnr.py
@@ -5,17 +5,13 @@ from warnings import warn
 import numpy as np
 from sklearn.base import RegressorMixin
 
-from squlearn.expectation_operator import ExpectationOperatorBase
-from squlearn.feature_map import FeatureMapBase
-from squlearn.optimizers import OptimizerBase, SGDMixin
-
 from .base_qnn import BaseQNN
 from .loss import LossBase
 from .training import solve_minibatch, regression
 
-from ..expectation_operator import ExpectationOperatorBase
-from ..feature_map import FeatureMapBase
-from ..optimizers import OptimizerBase
+from ..expectation_operator.expectation_operator_base import ExpectationOperatorBase
+from ..feature_map.feature_map_base import FeatureMapBase
+from ..optimizers.optimizer_base import OptimizerBase, SGDMixin
 from ..util import Executor
